requirements.txt
flask
selenium
webdriver-manager
pycryptodome
pytest
bandit
semgrep